SOLR-18109: move doStandardDebug/doStandardQueryDebug/doStandardResultsDebug/doSimpleQuery to DebugComponent
Resolves the file's own TODO: 'resolve duplicated code with DebugComponent.
Perhaps it should be added to doStandardDebug?' -- these 4 methods were
@Deprecated with 'move to DebugComponent' since SOLR-17561, but never moved.
Only 3 external call sites (2 in DebugComponent itself, 1 in
MoreLikeThisHandler). Widened SolrPluginUtils.explanationsToStrings from
private to public -- the only cross-package dependency the move needed,
since it wasn't itself deprecated and stays a general-purpose utility.
